Apply perf v4 and watcher fixes

+++ src/main/java/net/minecraft/entity/titan/EntityGargoyleTitan.java
@@ -141,7 +141,7 @@ implements IAnimatedEntity {
     @Override
     protected void entityInit() {
         super.entityInit();
-        this.dataWatcher.addObject(16, (Object)Byte.valueOf((byte)0));
+        this.dataWatcher.addObject(16, (Object)0);
     }
 
     @Override
